trips_specを細かく分ける
